adpcm: consolidate common ima implmenation details
